Add blocking scaling executor, move inner classes to top level files.
