+++ core/src/main/java/com/orientechnologies/common/thread/SingletonNamedThreadFactory.java
@@ -1,0 +1,16 @@
+package com.orientechnologies.common.thread;
+
+final class SingletonNamedThreadFactory extends BaseThreadFactory {
+
+  private final String name;
+
+  SingletonNamedThreadFactory(final String name, ThreadGroup parentThreadGroup) {
+    super(parentThreadGroup);
+    this.name = name;
+  }
+
+  @Override
+  protected String nextThreadName() {
+    return name;
+  }
+}
